Move navigation logic into NavigationBarController
This eliminates the MathController completely.
